# openapi.yaml
openapi: 3.0.3
info:
  title: Tasks API
  description: Kleine Notizen/Aufgaben-API mit E-Mail/Passwort-Auth, Pagination und Suche.
  version: 0.1.0
servers:
  - url: /
paths:
  /health:
    get:
      summary: Liveness-Check
      security: []
      responses:
        '200':
          description: Server läuft
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
  /auth/register:
    post:
      summary: Neuen Nutzer registrieren
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Credentials'
      responses:
        '201':
          description: Nutzer wurde angelegt
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '409':
          description: E-Mail ist bereits registriert
  /auth/login:
    post:
      summary: Einloggen und JWT erhalten
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Credentials'
      responses:
        '200':
          description: Login erfolgreich
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
        '401':
          description: E-Mail oder Passwort falsch
  /notes:
    get:
      summary: Eigene Notizen auflisten (paginiert, optional durchsuchbar)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Seite von Notizen
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotePage'
    post:
      summary: Neue Notiz anlegen
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NoteInput'
      responses:
        '201':
          description: Notiz wurde angelegt
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
  /notes/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    get:
      summary: Einzelne Notiz abrufen
      responses:
        '200':
          description: Notiz
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
        '404':
          description: Nicht gefunden
    patch:
      summary: Notiz teilweise aktualisieren
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NoteInput'
      responses:
        '200':
          description: Aktualisierte Notiz
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
        '404':
          description: Nicht gefunden
    delete:
      summary: Notiz löschen
      responses:
        '204':
          description: Gelöscht
        '404':
          description: Nicht gefunden
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Credentials:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        created_at:
          type: string
          format: date-time
    NoteInput:
      type: object
      properties:
        title:
          type: string
          maxLength: 200
        body:
          type: string
          maxLength: 10000
    Note:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        title:
          type: string
        body:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    NotePage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Note'
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
security:
  - bearerAuth: []